chatbot-openai/requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit